Fixed a type error in Process.java
The pid field was being stored as a long, but the posix.kill() method
expects an int pid.  Performed the cast-to-int in both kill() and
killProcessGroup().

+++ src/main/java/jnr/process/Process.java
@@ -81,7 +81,7 @@ public class Process {
     }
 
     public int kill(Signal sig) {
-        return posix.kill(pid, sig.intValue());
+        return posix.kill((int)pid, sig.intValue());
     }
 
     public int killProcessGroup() {
@@ -89,7 +89,7 @@ public class Process {
     }
 
     public int killProcessGroup(Signal sig) {
-        return posix.kill(-pid, sig.intValue());
+        return posix.kill(-(int)pid, sig.intValue());
     }
 
     public long exitValue() {
